@@ -8,7 +8,7 @@
     "start": "webpack serve --host 0.0.0.0 --mode production",
     "build:dist": "webpack --mode production",
     "build:watch": "webpack --mode production --watch",
-    "build": "npm i --no-optional --no-shrinkwrap --no-package-lock && npm run build:dist",
+    "build": "npm i --omit=optional --no-package-lock && npm run build:dist",
     "doc": "jsdoc -R README.md -P package.json -d doc -r -c jsdoc.config.json frontend",
     "test": "jest --passWithNoTests",
     "test:watch": "npm test -- --watch",
@@ -24,53 +24,55 @@
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/preset-env": "^7.15.4",
     "@babel/preset-react": "^7.14.5",
-    "@typescript-eslint/eslint-plugin": "^4.30.0",
-    "@typescript-eslint/parser": "^4.30.0",
-    "babel-jest": "^27.0.6",
-    "babel-loader": "^8.2.2",
+    "@typescript-eslint/eslint-plugin": "^5.54.0",
+    "@typescript-eslint/parser": "^5.54.0",
+    "babel-jest": "^29.4.3",
+    "babel-loader": "^9.1.2",
     "clean-webpack-plugin": "^4.0.0",
-    "copy-webpack-plugin": "^9.0.1",
+    "copy-webpack-plugin": "^11.0.0",
     "copyfiles": "^2.4.1",
     "css-loader": "^6.2.0",
-    "eslint": "^7.32.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-import": "2.24.2",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.25.1",
+    "eslint": "^8.35.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-react": "7.32.2",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-webpack-plugin": "3.0.1",
+    "eslint-webpack-plugin": "4.0.0",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.3.2",
-    "jest": "^27.0.6",
+    "jest": "^29.4.3",
     "jest-html-reporter": "^3.4.1",
-    "jsdoc": "3.6.7",
-    "react-test-renderer": "^17.0.2",
-    "rimraf": "^3.0.2",
+    "jsdoc": "4.0.2",
+    "react-test-renderer": "^18.2.0",
+    "rimraf": "^4.1.2",
     "style-loader": "^3.2.1",
     "webpack": "^5.52.0",
-    "webpack-cli": "^4.8.0",
+    "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.1.0"
   },
   "dependencies": {
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
     "@fontsource/roboto": "^4.5.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@mui/icons-material": "^5.11.9",
+    "@mui/lab": "^5.0.0-alpha.120",
+    "@mui/material": "^5.11.10",
     "@reduxjs/toolkit": "^1.6.1",
     "@types/leaflet": "^1.7.4",
-    "axios": "^0.21.3",
-    "i18next": "^20.6.0",
-    "i18next-browser-languagedetector": "^6.1.2",
-    "i18next-http-backend": "^1.3.0",
+    "axios": "^1.3.4",
+    "i18next": "^22.4.10",
+    "i18next-browser-languagedetector": "^7.0.1",
+    "i18next-http-backend": "^2.1.1",
     "leaflet": "^1.7.1",
     "loglevel": "^1.7.1",
     "material-ui-flags": "^1.2.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.11.4",
-    "react-leaflet": "^3.2.1",
-    "react-redux": "^7.2.5",
-    "react-router-dom": "^5.3.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^12.2.0",
+    "react-leaflet": "^4.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.8.2",
     "redux-logger": "^3.0.6",
     "regenerator-runtime": "^0.13.9",
     "yargs": "^17.1.1"
